moved model and fixed warning
